Move constant resolution inside the lexer.
This makes constants eligible for constant-folding.
It also fixes Kyromyr/Kyromyr.github.io#25, a bug where constants could
be assigned to.

An attempt was made to also move label resoultion in the same manner,
but because labels can be used on lines before they are defined, this
does not work.