Fix issue with string concat of number constants.
Fixes d0sboots/perfect-tower#6

The issue comes during constant folding, i.e. when concatenating two
constants. If the left one is a number, the result would be taken as a
number instead of a string.
This was rarely/never encountered, because concat is left-associative,
and you rarely start a concat chain with a number.